requirements.txt
pandas
fastapi[standard]
pinecone 
streamlit
langchain_huggingface 
langchain
langchain-community
pymupdf
Pillow
IPython
transformers
requests
matplotlib 
tourch